requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
python-multipart>=0.0.9
sse-starlette>=1.8
python-dotenv>=1.0
playwright>=1.40
pypdf>=4.0
pdf2image>=1.17
opencv-python>=4.9
numpy>=1.26
pillow>=10.0
sentry-sdk>=2.0
